Commit 5
- Creo URLs amigables de todos los enlaces.
-Traslado comprobar imagen del controler al post.
-Mejoro código y funcionalidades

+++ models/post.php
@@ -61,22 +61,55 @@ class Post {
         $req->execute();
     }
 
+    public static function comprobarImagen($nombre_img, $tipo, $tamano) {
+
+        //Si existe imagen y tiene un tamaño correcto
+        if (($nombre_img == !NULL) && ($_FILES['imagen']['size'] <= 900000000000)) {
+            //indicamos los formatos que permitimos subir a nuestro servidor
+            if (($_FILES["imagen"]["type"] == "image/gif") || ($_FILES["imagen"]["type"] == "image/jpeg") || ($_FILES["imagen"]["type"] == "image/jpg") || ($_FILES["imagen"]["type"] == "image/png")) {
+                // Ruta donde se guardarán las imágenes que subamos
+                $directorio = $_SERVER['DOCUMENT_ROOT'] . "/blog_php_mvc/uploads/";
+                // Muevo la imagen desde el directorio temporal a nuestra ruta indicada anteriormente
+                move_uploaded_file($_FILES['imagen']['tmp_name'], $directorio . $nombre_img);
+            } else {
+                //si no cumple con el formato
+                echo "No se puede subir una imagen con ese formato ";
+            }
+        } else {
+            //si existe la variable pero se pasa del tamaño permitido
+            if ($nombre_img == !NULL)
+                echo "La imagen es demasiado grande ";
+        }
+
+        return $nombre_img;
+    }
+
     // Definimos formato para la fecha y hacemos la sentencia update .
     public static function update($titulo, $autor, $contenido, $imagen, $id) {
         $db = Db::getInstance();
 
-        $timestamp = date('Y-m-d H:i:s');
-
-        $req = $db->prepare("UPDATE posts SET modified=:modified, titulo=:titulo, imagen=:imagen, author=:author, content =:content WHERE id=:id");
-        $req->execute(array('modified' => $timestamp, 'titulo' => $titulo, 'imagen' => $imagen, 'author' => $autor, 'content' => $contenido, 'id' => $id));
+        
+        if ($imagen != "") {
+            $timestamp = date('Y-m-d H:i:s');
+            $req = $db->prepare("UPDATE posts SET modified=:modified, titulo=:titulo, imagen=:imagen, author=:author, content =:content WHERE id=:id");
+            $req->execute(array('modified' => $timestamp, 'titulo' => $titulo, 'imagen' => $imagen, 'author' => $autor, 'content' => $contenido, 'id' => $id));
+        } else {
+            $timestamp = date('Y-m-d H:i:s');
+            $req = $db->prepare("UPDATE posts SET modified=:modified, titulo=:titulo, author=:author, content =:content WHERE id=:id");
+            $req->execute(array('modified' => $timestamp, 'titulo' => $titulo, 'author' => $autor, 'content' => $contenido, 'id' => $id));
+        }
     }
 
     //Utilizando la id que enviamos por parametro para borrar el post
     public static function borrar($id) {
         $db = Db::getInstance();
 
-        $req = $db->prepare("DELETE FROM posts WHERE id=:id");
-        $req->execute(array('id' => $id));
+        if (isset($_GET['id'])) {
+            $id = $_GET['id'];
+
+            $req = $db->prepare("DELETE FROM posts WHERE id=:id");
+            $req->execute(array('id' => $id));
+        }
     }
 
     public function readAll($from_record_num, $records_per_page) {
